# static/openapi.yaml
openapi: "3.0.3"
info:
  title: AiPayGen API
  description: |
    AI-powered API platform with 244+ tools. Pay-per-call with API keys (Bearer apk_xxx),
    x402 V2 USDC micropayments (multi-chain: Base, Solana, Stellar), or use the free tier (1 call/day per IP).
    Supports autonomous agent micropayments — no API key or signup required.

    Base URL: https://api.aipaygen.com
  version: "1.8.1"
  contact:
    email: hello@aipaygen.com
    url: https://aipaygen.com
  license:
    name: Proprietary
servers:
  - url: https://api.aipaygen.com
    description: Production
security:
  - BearerApiKey: []
  - AgentJWT: []
  - {}
components:
  securitySchemes:
    BearerApiKey:
      type: http
      scheme: bearer
      description: "API key with apk_ prefix. Obtain via POST /auth/generate-key or POST /credits/buy."
    AgentJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Agent JWT obtained via /agents/challenge + /agents/verify wallet auth flow."
...
paths:
...
  /sell/register:
    post:
      tags: [Seller]
      summary: Register a new seller API
      security:
        - BearerApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, slug, base_url]
              properties:
                name:
                  type: string
                slug:
                  type: string
                  description: "URL-safe identifier for your API"
                description:
                  type: string
                base_url:
                  type: string
                  description: "HTTPS base URL of your API"
                routes:
                  type: array
                  items:
                    type: object
                    properties:
                      path:
                        type: string
                      method:
                        type: string
                      price_usd:
                        type: number
                      description:
                        type: string
                seller_wallet:
                  type: string
                preferred_chain:
                  type: string
                  default: base
                category:
                  type: string
                  default: general
                escrow_enabled:
                  type: boolean
                  default: false
      responses:
        "201":
          description: Seller API registered
        "400":
          description: Missing required fields
        "401":
          description: API key required
  /sell/dashboard:
    get:
      tags: [Seller]
      summary: Seller analytics dashboard
      security:
        - BearerApiKey: []
      responses:
        "200":
          description: Dashboard data with revenue and call stats
        "401":
          description: API key required

  # ── Sessions ──────────────────────────────────────────────────────────
...